refactor: organize with standard structure
Also enable a module import
