+++ src/ridewithgps_to_cuesheet/__init__.py
@@ -1,0 +1,6 @@
+"""RideWithGPS to Cuesheet converter package."""
+
+from .ridewithgps import generate_excel, read_csv_to_array, format_array, argsobject
+
+__version__ = "0.4.0"
+__all__ = ["generate_excel", "read_csv_to_array", "format_array", "argsobject"]
